ファイル名の変更と読み取り結果の取り消し実装
