+++ web/src/App.tsx
@@ -3,7 +3,7 @@ import './App.css';
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
 
-import TicketQRReader from './components/TicketQRReader';
+import Ticket from './components/Ticket';
 import GlobalHeader from './layouts/GlobalHeader';
 
 function App() {
@@ -12,7 +12,7 @@ function App() {
       <Container disableGutters maxWidth="md" sx={{height: "100%", display: "flex", flexDirection: "column"}}>
         <GlobalHeader />
         <Box sx={{flexGrow: 1, bgcolor: "background.paper", overflow: "scroll"}}>
-          <TicketQRReader />
+          <Ticket />
         </Box>
       </Container>
     </div>
